temp checkin, gonna mess stuff up

+++ Nez/Nez-PCL/UI/Widgets/SelectBox.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Microsoft.Xna.Framework;
 using Nez.BitmapFonts;
 
+
 namespace Nez.UI
 {
 	public class SelectBox<T> : Element, IInputListener where T : class
@@ -90,7 +91,7 @@ namespace Nez.UI
 			var fontColor = _isDisabled ? style.disabledFontColor : style.fontColor;
 
 			var color = getColor();
-			color = new Color( color, color.A * parentAlpha );
+			color = new Color( color, (int)(color.A * parentAlpha) );
 			float x = getX();
 			float y = getY();
 			float width = getWidth();
@@ -115,7 +116,7 @@ namespace Nez.UI
 					y += (int)( height / 2 + font.lineHeight / 2 );
 				}
 
-				fontColor = new Color( fontColor, fontColor.A * parentAlpha );
+				fontColor = new Color( fontColor, (int)(fontColor.A * parentAlpha) );
 				graphics.batcher.drawString( font, str, new Vector2( x, y ), fontColor );
 			}
 		}
